resume_screening_backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.6
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
filelock==3.9.0
joblib==1.4.2
lxml==5.3.0
nltk==3.9.1
numpy==2.1.3
pdfminer.six==20240706
platformdirs==2.6.2
pycparser==2.22
PyPDF2==3.0.1
python-docx==1.1.2
regex==2024.11.6
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.4.3
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2022.7
virtualenv==20.17.1